Move: 재설계된 패키지 구조로 기존 작업 파일들 이동
