+++ src/main/java/trendravel/photoravel_be/commom/results/OnlyResultDto.java
@@ -1,0 +1,14 @@
+package trendravel.photoravel_be.commom.results;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import trendravel.photoravel_be.commom.results.messages.ResultInfo;
+
+
+@Data
+@NoArgsConstructor
+public class OnlyResultDto{
+
+    private ResultInfo result;
+
+}
