+++ src/main/java/trendravel/photoravel_be/db/respository/LocationRepository.java
@@ -1,0 +1,8 @@
+package trendravel.photoravel_be.db.respository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import trendravel.photoravel_be.db.location.Location;
+
+public interface LocationRepository extends JpaRepository<Location, Long> {
+
+}
