--- src/main/java/trendravel/photoravel_be/dto/response/results/OnlyResultDto.java
@@ -1,15 +1,0 @@
-package trendravel.photoravel_be.dto.response.results;
-
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import trendravel.photoravel_be.dto.response.messages.ResultInfo;
-
-
-@Data
-@NoArgsConstructor
-public class OnlyResultDto{
-
-    private ResultInfo result;
-
-}
