--- src/main/java/trendravel/photoravel_be/repository/ReviewRepository.java
@@ -1,8 +1,0 @@
-package trendravel.photoravel_be.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import trendravel.photoravel_be.entity.Review;
-
-public interface ReviewRepository extends JpaRepository<Review, Long>{
-
-}
